Fix comment length overflow.

+++ src/mz_compat.h
@@ -176,7 +176,7 @@ typedef voidp unzFile;
 typedef struct unz_global_info64_s
 {
     uint64_t number_entry;          // total number of entries in the central dir on this disk 
-    uint32_t number_disk_with_CD;   // number the the disk with central dir, used for spanning ZIP
+    uint64_t number_disk_with_CD;   // number the the disk with central dir, used for spanning ZIP
     uint16_t size_comment;          // size of the global comment of the zipfile 
 } unz_global_info64;
